Extract benchmark receipts into SRP microcrate
