+++ Cargo.toml
@@ -67,6 +67,7 @@ members = [
     "crates/bitnet-rocm",          # HIP compute kernels for AMD ROCm backend",
   "crates/bitnet-vulkan",        # Vulkan GLSL compute shaders for GPU inference
   "crates/bitnet-wgpu",          # WebGPU compute backend for cross-platform GPU inference
+  "crates/bitnet-bench-receipts", # SRP: benchmark receipt model + JSONL persistence
   "crates/bitnet-wgpu-bench",    # Benchmarking and profiling harness for wgpu compute kernels
   "crossval",
   "tests",
