+++ crates/bitnet-bench-receipts/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "bitnet-bench-receipts"
+version = "0.1.0"
+edition = "2024"
+rust-version = "1.92.0"
+description = "Receipt model and JSONL store for benchmark result tracking"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+thiserror = "2"
+
+[dev-dependencies]
+tempfile = "3"
