+++ crates/bitnet-wgpu-bench/Cargo.toml
@@ -7,10 +7,7 @@ description = "Benchmarking and profiling harness for wgpu compute kernels"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-thiserror = "2"
+bitnet-bench-receipts = { path = "../bitnet-bench-receipts" }
 
 [dev-dependencies]
 proptest = "1"
-tempfile = "3"
